Keep width and use the suggesting behavior from editing notes in TransactionEdit.tsx
