+++ packages/desktop-client/src/components/mobile/transactions/TransactionEdit.tsx
@@ -64,7 +64,6 @@ import type {
   PayeeEntity,
   TransactionEntity,
 } from '@actual-app/core/types/models';
-import { css } from '@emotion/css';
 import {
   format as formatDate,
   isValid as isValidDate,
@@ -72,6 +71,7 @@ import {
   parseISO,
 } from 'date-fns';
 
+import { NoteTagAutocomplete } from '#components/autocomplete/NoteTagAutocomplete';
 import { MobileBackButton } from '#components/mobile/MobileBackButton';
 import {
   FieldLabel,
@@ -84,7 +84,6 @@ import { MobilePageHeader, Page } from '#components/Page';
 import { shouldApplyRuleChange } from '#components/transactions/table/utils';
 import { useAccounts } from '#hooks/useAccounts';
 import { useCategories } from '#hooks/useCategories';
-import { useCurrentWordRange } from '#hooks/useCurrentWordRange';
 import { useCursorPosition } from '#hooks/useCursorPosition';
 import { useDateFormat } from '#hooks/useDateFormat';
 import { useInputRefValue } from '#hooks/useInputRefValue';
@@ -98,8 +97,6 @@ import {
   useSingleActiveEditForm,
 } from '#hooks/useSingleActiveEditForm';
 import { useSyncedPref } from '#hooks/useSyncedPref';
-import { useTagCSS } from '#hooks/useTagCSS';
-import { useFilteredTags } from '#hooks/useTags';
 import { pushModal } from '#modals/modalsSlice';
 import { addNotification } from '#notifications/notificationsSlice';
 import { useSavePayeeLocationMutation } from '#payees';
@@ -1590,151 +1587,6 @@ function NoteInsertHashButton({
     >
       <SvgHash width={17} height={17} />
     </Button>
-  );
-}
-
-function NoteTagAutocomplete({
-  inputRef,
-}: {
-  inputRef: RefObject<HTMLInputElement | null>;
-}) {
-  const dispatch = useDispatch();
-  // Yes, there is a lot of ref usages in this component. Here's the motivation
-  // 1. This component purely modifies HTML Input state, app state is handled elsewhere
-  // 2. This component deals with cursor state, which is not easily accessible through regular React code
-  // 3. Child transaction notes (transaction.notes) does not update until blur, so we have to use input state
-  // 4. Given we are already using inputRef in multiple locations, I elected to simplify the props to just the ref and use HTML/JS events
-
-  const [note, setNote] = useInputRefValue(inputRef);
-
-  const [cursorPosition] = useCursorPosition(inputRef);
-  const [startIdx, endIdx] = useCurrentWordRange(note, cursorPosition);
-  const currentWord = note.slice(startIdx, endIdx);
-  const currentWordNoHash = currentWord.replace(/^#+/, '');
-  const { data: filteredTags, refetch } = useFilteredTags(currentWord, true);
-  const showNewTag =
-    currentWord.startsWith('#') &&
-    currentWordNoHash &&
-    !filteredTags.some(tag => tag.tag === currentWordNoHash);
-
-  const getTagCSS = useTagCSS({ ellipsis: true });
-
-  function handleSelect(tag: string) {
-    if (!inputRef.current) return;
-    const newValue =
-      note.slice(0, startIdx) + '#' + tag + ' ' + note.slice(endIdx);
-    setNote(newValue);
-    const newPos = startIdx + tag.length + 2;
-
-    inputRef.current.setSelectionRange(newPos, newPos);
-    document.dispatchEvent(new Event('selectionchange'));
-  }
-
-  async function handleCreate(tag: string) {
-    if (!inputRef.current) return;
-    try {
-      await send('tags-create', { tag });
-      void refetch();
-      handleSelect(tag);
-    } catch (e) {
-      dispatch(
-        addNotification({
-          notification: {
-            type: 'error',
-            message: 'Failed to add tag, check logs',
-          },
-        }),
-      );
-      console.trace(e);
-    }
-  }
-
-  const hideScrollbar = css({
-    'scrollbar-width': 'none',
-    '-ms-overflow-style': 'none',
-    '&::-webkit-scrollbar': {
-      display: 'none',
-    },
-  });
-
-  return (
-    <View
-      style={{
-        width: '100%',
-        padding: '4px 8px 4px 8px',
-        borderRadius: 30,
-        overflowX: 'auto',
-        height: filteredTags.length || showNewTag ? 30 : 0,
-        transitionProperty: 'height',
-        transitionDuration: '100ms',
-      }}
-      className={hideScrollbar}
-    >
-      <View
-        style={{
-          display: 'flex',
-          flexDirection: 'row',
-          alignItems: 'end',
-          flexWrap: 'nowrap',
-          gap: 4,
-          paddingRight: 8,
-        }}
-      >
-        {filteredTags.map(tag => (
-          <div key={tag.id}>
-            <button
-              type="button"
-              style={{
-                border: 'none',
-                height: 22,
-                maxWidth: '50dvw',
-              }}
-              className={getTagCSS(tag.tag)}
-              onMouseDown={e => e.preventDefault()} // stops input from losing focus
-              onClick={() => handleSelect(tag.tag)}
-            >
-              #{tag.tag}
-            </button>
-          </div>
-        ))}
-        {showNewTag && (
-          <button
-            type="button"
-            style={{
-              padding: '1px 1px 1px 9px',
-              borderRadius: 12,
-              borderWidth: 0,
-              backgroundColor: theme.noticeBackground,
-              color: theme.noticeTextDark,
-              display: 'flex',
-              alignItems: 'center',
-              flexWrap: 'nowrap',
-              gap: 4,
-            }}
-            onMouseDown={e => e.preventDefault()} // stops input from losing focus
-            onClick={() => handleCreate(currentWordNoHash)}
-          >
-            <SvgAdd height={8} width={8} />
-            <span style={{ whiteSpace: 'nowrap' }}>
-              <Trans>Create tag</Trans>
-            </span>
-            <div
-              style={{
-                borderWidth: 0,
-                height: 20,
-                maxWidth: '50dvw',
-                overflow: 'hidden',
-                textOverflow: 'ellipsis',
-                display: 'inline-block',
-              }}
-              className={getTagCSS('')}
-            >
-              #{currentWordNoHash}
-            </div>
-          </button>
-        )}
-      </View>
-    </View>
   );
 }
