+++ upcoming-release-notes/add-tags-to-bulk-edit-notes.md
@@ -3,4 +3,4 @@ category: Enhancements
 authors: [Dreptschar]
 ---
 
-Show existing tags below the notes field when bulk editing transactions on mobile, so you can tap to add them
+Suggest matching tags as you type # in the notes field when bulk editing transaction notes
